Move print to logger to remtor.py
